requirements.txt
streamlit
requests
python-dotenv
pyyaml
pandas
numpy
